+++ .idea/fileTemplates/GUIScreenController.java
@@ -1,11 +1,11 @@
-package client.GUI.Forms;
+package ${PACKAGE_NAME};
 
 import client.GUI.CPSClientGUI;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 
-public class ManageSubscription {
+public class ${NAME} {
 
     @FXML
     private Button btnBack;