+++ src/main/client/Changelog.md
@@ -1,5 +1,10 @@
 # 14.01.2018
 * **12:11** - Fixing some UI bugs with deleting cars.
+* **15:15** - Added Email validation in registration/login. `u` is still accepted though.
+* **16:17** - Fixed bug with field highlighting. Started working on Subscription screen.
+* Fixed bug with cells not showing all information.
+* **18:05** - Finished the "Add Subscription" GUI and validation. Still WIP: The client/server flow.
+* **20:15** - Finished the entire "Subscription" GUI and client-side logic.
 
 # 13.01.2018
 ## General